utils: Stop fixing AUTHENTICATION_BACKENDS unnecessarily.
If you're writing unit tests that test various options for
AUTHENTICATION_BACKENDS using e.g.
@override_settings(AUTHENTICATION_BACKENDS=...) or `with
self.settings(...)`, that works great in Django in general.

However, the first such test that interacts with python-social-auth
enough to cause utils.py to be imported will fix the value of
AUTHENTICATION_BACKENDS used by social-auth for the rest of the
process' lifetime.

As a result, if one first accesses it with a smaller/restrictive list
of backends, and then tries to use it with a larger list later, one
gets a 404 "Backend Not Found" error in those later accesses. So it
manifested for us as unit tests passing or failing depending on
whether our one test that sets a short list of backends ran early or
late.

The fix is simply to access settings.AUTHENTICATION_BACKENDS directly
when it needs it, rather than providing a short alias at the top of
the utils.py file.

Fixes #161.